Update addon via `ember init`
