+++ package.json
@@ -4,17 +4,14 @@
   "author": "James A. Rosen <james.a.rosen@gmail.com>",
   "version": "1.3.4",
   "licence": "APLv2 / MIT",
-  "contributors": [
-    "Miguel Camba <miguel.camba@gmail.com>"
-  ],
   "directories": {
     "doc": "doc",
     "test": "tests"
   },
   "scripts": {
-    "start": "ember server",
     "build": "ember build",
-    "test": "ember test"
+    "start": "ember server",
+    "test": "ember try:testall"
   },
   "repository": {
     "type": "git",
@@ -23,34 +20,38 @@
   "engines": {
     "node": ">= 0.10.0"
   },
+  "author": "Miguel Camba <miguel.camba@gmail.com>",
+  "license": "MIT",
   "devDependencies": {
-    "broccoli-asset-rev": "^2.0.2",
+    "broccoli-asset-rev": "^2.1.2",
     "broccoli-defeatureify": "^1.0.0",
     "broccoli-dist-es6-module": "0.2.1",
-    "broccoli-ember-hbs-template-compiler": "^1.6.1",
     "broccoli-funnel": "^0.2.3",
     "broccoli-merge-trees": "^0.2.1",
     "broccoli-uglify-js": "^0.1.3",
-    "ember-cli": "^0.2.7",
-    "ember-cli-app-version": "0.3.3",
+    "ember-cli": "1.13.8",
+    "ember-cli-app-version": "^1.0.0",
     "ember-cli-content-security-policy": "0.4.0",
-    "ember-cli-htmlbars": "~0.7.9",
-    "ember-cli-ic-ajax": "~0.1.2",
-    "ember-cli-inject-live-reload": "^1.3.0",
-    "ember-cli-qunit": "~0.3.14",
-    "ember-cli-uglify": "1.0.1",
+    "ember-cli-dependency-checker": "^1.1.0",
+    "ember-cli-htmlbars": "^1.0.0",
+    "ember-cli-htmlbars-inline-precompile": "^0.3.1",
+    "ember-cli-inject-live-reload": "^1.3.1",
+    "ember-cli-qunit": "^1.0.1",
+    "ember-cli-release": "0.2.3",
+    "ember-cli-uglify": "^1.2.0",
     "ember-cli-yuidoc": "~0.7.0",
-    "ember-disable-prototype-extensions": "1.0.1",
-    "ember-export-application-global": "^1.0.2",
+    "ember-disable-prototype-extensions": "^1.0.0",
+    "ember-export-application-global": "^1.0.4",
+    "ember-resolver": "^2.0.2",
     "ember-publisher": "0.0.7",
-    "ember-try": "0.0.7"
+    "ember-try": "~0.0.8"
   },
   "keywords": [
     "ember",
     "ember-addon"
   ],
   "dependencies": {
-    "ember-cli-babel": "^5.0.0"
+    "ember-cli-babel": "^5.1.3"
   },
   "ember-addon": {
     "configPath": "tests/dummy/config"
