+++ tests/index.html
@@ -24,7 +24,7 @@
     <script src="assets/vendor.js"></script>
     <script src="assets/test-support.js"></script>
     <script src="assets/dummy.js"></script>
-    <script src="testem.js"></script>
+    <script src="testem.js" integrity=""></script>
     <script src="assets/test-loader.js"></script>
 
     {{content-for 'body-footer'}}
